+++ src/test/java/roomescape/controller/AdminControllerTest.java
@@ -1,0 +1,25 @@
+package roomescape.controller;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+@DisplayNameGeneration(ReplaceUnderscores.class)
+class AdminControllerTest {
+    @Test
+    void 관리_페이지_접근시_정상응답() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+
+}
