+++ src/test/java/roomescape/fake/FakeReservaionDao.java
@@ -1,4 +1,4 @@
-package roomescape.Fake;
+package roomescape.fake;
 
 import java.util.ArrayList;
 import java.util.List;
