+++ src/test/java/roomescape/fake/FakeReservationTimeDao.java
@@ -1,13 +1,11 @@
-package roomescape.Fake;
+package roomescape.fake;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
 import roomescape.dao.ReservationTimeDao;
-import roomescape.dto.ReservationTimeResponse;
 import roomescape.model.ReservationTime;
 
 public class FakeReservationTimeDao implements ReservationTimeDao {
