+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -9,8 +9,8 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.web.server.ResponseStatusException;
-import roomescape.Fake.FakeReservaionDao;
-import roomescape.Fake.FakeReservationTimeDao;
+import roomescape.fake.FakeReservaionDao;
+import roomescape.fake.FakeReservationTimeDao;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dto.ReservationRequest;
